Move CurlQueryOp to separate source file
